Move GraphPrinter into myia/ir/visualization.
