+++ myia/ir/visualization/__init__.py
@@ -1,0 +1,1 @@
+from .graph_printer import GraphPrinter
